requirements.txt
discord.py
jisukaku